Separate Redis connections into cache and session
