+++ src/AppBundle.php
@@ -6,7 +6,8 @@ use eLife\Journal\DependencyInjection\GuzzleMiddlewarePass;
 use eLife\Journal\DependencyInjection\HttpCachePass;
 use eLife\Journal\DependencyInjection\MonologStacktracesPass;
 use eLife\Journal\DependencyInjection\OAuthClientPass;
-use eLife\Journal\DependencyInjection\RedisPass;
+use eLife\Journal\DependencyInjection\RedisCachePass;
+use eLife\Journal\DependencyInjection\RedisSessionsPass;
 use eLife\Journal\DependencyInjection\ViewModelConverterPass;
 use Symfony\Component\DependencyInjection\Compiler\PassConfig;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -20,7 +21,8 @@ final class AppBundle extends Bundle
         $container->addCompilerPass(new HttpCachePass(), PassConfig::TYPE_BEFORE_OPTIMIZATION, 33);
         $container->addCompilerPass(new MonologStacktracesPass());
         $container->addCompilerPass(new OAuthClientPass());
-        $container->addCompilerPass(new RedisPass(), PassConfig::TYPE_BEFORE_OPTIMIZATION, 34);
+        $container->addCompilerPass(new RedisCachePass(), PassConfig::TYPE_BEFORE_OPTIMIZATION, 34);
+        $container->addCompilerPass(new RedisSessionsPass(), PassConfig::TYPE_BEFORE_OPTIMIZATION, 34);
         $container->addCompilerPass(new ViewModelConverterPass());
     }
 }
